+++ CHANGELOG.md
@@ -32,6 +32,9 @@ and this project adheres to
   simply set `model-id` to either 'random-electra-sequence-clf or
   'random-electra-token-clf'. The randomly initialised XLM-RoBERTa-base model
   is still available by replacing 'electra' with 'roberta'.
+- Added `--raise-error-on-invalid-model` flag which raises an exception if an invalid
+  model is specified. By default this is off, meaning that it simply skips the model if
+  it is invalid.
 
 ### Fixed
 - Now disables tokenizer progress bars properly, using the